+++ data-raw/IPCC_CO2_concentration.R
@@ -1,0 +1,3 @@
+library(usethis)
+cCO2_coefficients = read.csv("inst/extdata/IPCC_CO2_concentration.csv")
+usethis::use_data(cCO2_coefficients)